move to new package (and move Validate too)
