+++ common/constants.go
@@ -69,14 +69,6 @@ const (
 )
 
 const (
-	// TimeoutFailureTypePrefix is the prefix for timeout failure types
-	// used in retry policy
-	// the actual failure type will be prefix + enums.TimeoutType.String()
-	// e.g. "TemporalTimeout:StartToClose" or "TemporalTimeout:Heartbeat"
-	TimeoutFailureTypePrefix = "TemporalTimeout:"
-)
-
-const (
 	// Limit for schedule notes field
 	ScheduleNotesSizeLimit = 1000
 )
