+++ common/retrypolicy/retry_policy.go
@@ -22,12 +22,16 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
-package dynamicconfig
+package retrypolicy
 
 import (
+	"fmt"
+	"strings"
 	"time"
 
 	commonpb "go.temporal.io/api/common/v1"
+	enumspb "go.temporal.io/api/enums/v1"
+	"go.temporal.io/api/serviceerror"
 	"google.golang.org/protobuf/types/known/durationpb"
 
 	"go.temporal.io/server/common/number"
@@ -35,6 +39,12 @@ import (
 )
 
 const (
+	// TimeoutFailureTypePrefix is the prefix for timeout failure types
+	// used in retry policy
+	// the actual failure type will be prefix + enums.TimeoutType.String()
+	// e.g. "TemporalTimeout:StartToClose" or "TemporalTimeout:Heartbeat"
+	TimeoutFailureTypePrefix = "TemporalTimeout:"
+
 	defaultInitialInterval            = time.Second
 	defaultMaximumIntervalCoefficient = 100.0
 	defaultBackoffCoefficient         = 2.0
@@ -46,7 +56,7 @@ const (
 	maximumAttemptsConfigKey            = "MaximumAttempts"
 )
 
-func GetDefaultRetryPolicyConfigOptions() map[string]any {
+func GetDefault() map[string]any {
 	return map[string]any{
 		initialIntervalInSecondsConfigKey:   int(defaultInitialInterval.Seconds()),
 		maximumIntervalCoefficientConfigKey: defaultMaximumIntervalCoefficient,
@@ -65,7 +75,7 @@ type DefaultRetrySettings struct {
 	MaximumAttempts            int32
 }
 
-func FromConfigToDefaultRetrySettings(options map[string]interface{}) DefaultRetrySettings {
+func FromConfigToDefault(options map[string]interface{}) DefaultRetrySettings {
 	defaultSettings := DefaultRetrySettings{
 		InitialInterval: defaultInitialInterval,
 		MaximumAttempts: defaultMaximumAttempts,
@@ -100,8 +110,8 @@ func FromConfigToDefaultRetrySettings(options map[string]interface{}) DefaultRet
 	return defaultSettings
 }
 
-// EnsureRetryPolicyDefaults ensures the policy subfields, if not explicitly set, are set to the specified defaults
-func EnsureRetryPolicyDefaults(originalPolicy *commonpb.RetryPolicy, defaultSettings DefaultRetrySettings) {
+// EnsureDefaults ensures the policy subfields, if not explicitly set, are set to the specified defaults
+func EnsureDefaults(originalPolicy *commonpb.RetryPolicy, defaultSettings DefaultRetrySettings) {
 	if originalPolicy.GetMaximumAttempts() == 0 {
 		originalPolicy.MaximumAttempts = defaultSettings.MaximumAttempts
 	}
@@ -117,4 +127,45 @@ func EnsureRetryPolicyDefaults(originalPolicy *commonpb.RetryPolicy, defaultSett
 	if originalPolicy.GetBackoffCoefficient() == 0 {
 		originalPolicy.BackoffCoefficient = defaultSettings.BackoffCoefficient
 	}
+}
+
+// Validate validates a retry policy
+func Validate(policy *commonpb.RetryPolicy) error {
+	if policy == nil {
+		// nil policy is valid which means no retry
+		return nil
+	}
+
+	if policy.GetMaximumAttempts() == 1 {
+		// One maximum attempt effectively disable retries. Validating the
+		// rest of the arguments is pointless
+		return nil
+	}
+	if timestamp.DurationValue(policy.GetInitialInterval()) < 0 {
+		return serviceerror.NewInvalidArgument("InitialInterval cannot be negative on retry policy.")
+	}
+	if policy.GetBackoffCoefficient() < 1 {
+		return serviceerror.NewInvalidArgument("BackoffCoefficient cannot be less than 1 on retry policy.")
+	}
+	if timestamp.DurationValue(policy.GetMaximumInterval()) < 0 {
+		return serviceerror.NewInvalidArgument("MaximumInterval cannot be negative on retry policy.")
+	}
+	if timestamp.DurationValue(policy.GetMaximumInterval()) > 0 && timestamp.DurationValue(policy.GetMaximumInterval()) < timestamp.DurationValue(policy.GetInitialInterval()) {
+		return serviceerror.NewInvalidArgument("MaximumInterval cannot be less than InitialInterval on retry policy.")
+	}
+	if policy.GetMaximumAttempts() < 0 {
+		return serviceerror.NewInvalidArgument("MaximumAttempts cannot be negative on retry policy.")
+	}
+
+	for _, nrt := range policy.NonRetryableErrorTypes {
+		if strings.HasPrefix(nrt, TimeoutFailureTypePrefix) {
+			timeoutTypeValue := nrt[len(TimeoutFailureTypePrefix):]
+			timeoutType, err := enumspb.TimeoutTypeFromString(timeoutTypeValue)
+			if err != nil || enumspb.TimeoutType(timeoutType) == enumspb.TIMEOUT_TYPE_UNSPECIFIED {
+				return serviceerror.NewInvalidArgument(fmt.Sprintf("Invalid timeout type value: %v.", timeoutTypeValue))
+			}
+		}
+	}
+
+	return nil
 }
