+++ common/retrypolicy/retry_policy_test.go
@@ -22,7 +22,7 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
-package dynamicconfig
+package retrypolicy
 
 import (
 	"testing"
@@ -30,6 +30,7 @@ import (
 
 	"github.com/stretchr/testify/assert"
 	commonpb "go.temporal.io/api/common/v1"
+	enumspb "go.temporal.io/api/enums/v1"
 	"google.golang.org/protobuf/types/known/durationpb"
 )
 
@@ -123,7 +124,7 @@ func TestEnsureRetryPolicyDefaults(t *testing.T) {
 
 	for _, tt := range testCases {
 		t.Run(tt.name, func(t *testing.T) {
-			EnsureRetryPolicyDefaults(tt.input, defaultRetrySettings)
+			EnsureDefaults(tt.input, defaultRetrySettings)
 			assert.Equal(t, tt.want, tt.input)
 		})
 	}
@@ -137,9 +138,126 @@ func Test_FromConfigToRetryPolicy(t *testing.T) {
 		maximumAttemptsConfigKey:            5,
 	}
 
-	defaultSettings := FromConfigToDefaultRetrySettings(options)
+	defaultSettings := FromConfigToDefault(options)
 	assert.Equal(t, 2*time.Second, defaultSettings.InitialInterval)
 	assert.Equal(t, 100.0, defaultSettings.MaximumIntervalCoefficient)
 	assert.Equal(t, 4.0, defaultSettings.BackoffCoefficient)
 	assert.Equal(t, int32(5), defaultSettings.MaximumAttempts)
+}
+
+func TestValidateRetryPolicy(t *testing.T) {
+	testCases := []struct {
+		name          string
+		input         *commonpb.RetryPolicy
+		wantErr       bool
+		wantErrString string
+	}{
+		{
+			name:          "nil policy is okay",
+			input:         nil,
+			wantErr:       false,
+			wantErrString: "",
+		},
+		{
+			name: "maxAttempts is 1, coefficient < 1",
+			input: &commonpb.RetryPolicy{
+				BackoffCoefficient: 0.5,
+				MaximumAttempts:    1,
+			},
+			wantErr:       false,
+			wantErrString: "",
+		},
+		{
+			name: "initial interval negative",
+			input: &commonpb.RetryPolicy{
+				InitialInterval: durationpb.New(-22 * time.Second),
+			},
+			wantErr:       true,
+			wantErrString: "InitialInterval cannot be negative on retry policy.",
+		},
+		{
+			name: "coefficient < 1",
+			input: &commonpb.RetryPolicy{
+				BackoffCoefficient: 0.8,
+			},
+			wantErr:       true,
+			wantErrString: "BackoffCoefficient cannot be less than 1 on retry policy.",
+		},
+		{
+			name: "maximum interval in seconds is negative",
+			input: &commonpb.RetryPolicy{
+				BackoffCoefficient: 2.0,
+				MaximumInterval:    durationpb.New(-2 * time.Second),
+			},
+			wantErr:       true,
+			wantErrString: "MaximumInterval cannot be negative on retry policy.",
+		},
+		{
+			name: "maximum interval in less than initial interval",
+			input: &commonpb.RetryPolicy{
+				BackoffCoefficient: 2.0,
+				MaximumInterval:    durationpb.New(5 * time.Second),
+				InitialInterval:    durationpb.New(10 * time.Second),
+			},
+			wantErr:       true,
+			wantErrString: "MaximumInterval cannot be less than InitialInterval on retry policy.",
+		},
+		{
+			name: "maximum attempts negative",
+			input: &commonpb.RetryPolicy{
+				BackoffCoefficient: 2.0,
+				MaximumAttempts:    -3,
+			},
+			wantErr:       true,
+			wantErrString: "MaximumAttempts cannot be negative on retry policy.",
+		},
+		{
+			name: "timeout nonretryable error - valid type",
+			input: &commonpb.RetryPolicy{
+				BackoffCoefficient: 1,
+				NonRetryableErrorTypes: []string{
+					TimeoutFailureTypePrefix + enumspb.TIMEOUT_TYPE_START_TO_CLOSE.String(),
+					TimeoutFailureTypePrefix + enumspb.TIMEOUT_TYPE_SCHEDULE_TO_START.String(),
+					TimeoutFailureTypePrefix + enumspb.TIMEOUT_TYPE_SCHEDULE_TO_CLOSE.String(),
+					TimeoutFailureTypePrefix + enumspb.TIMEOUT_TYPE_HEARTBEAT.String(),
+				},
+			},
+			wantErr:       false,
+			wantErrString: "",
+		},
+		{
+			name: "timeout nonretryable error - unspecified type",
+			input: &commonpb.RetryPolicy{
+				BackoffCoefficient: 1,
+				NonRetryableErrorTypes: []string{
+					TimeoutFailureTypePrefix + enumspb.TIMEOUT_TYPE_UNSPECIFIED.String(),
+				},
+			},
+			wantErr:       true,
+			wantErrString: "Invalid timeout type value: Unspecified.",
+		},
+		{
+			name: "timeout nonretryable error - unknown type",
+			input: &commonpb.RetryPolicy{
+				BackoffCoefficient: 1,
+				NonRetryableErrorTypes: []string{
+					TimeoutFailureTypePrefix + "unknown",
+				},
+			},
+			wantErr:       true,
+			wantErrString: "Invalid timeout type value: unknown.",
+		},
+	}
+
+	for _, tt := range testCases {
+		t.Run(tt.name, func(t *testing.T) {
+			err := Validate(tt.input)
+			if tt.wantErr {
+				assert.NotNil(t, err, "expected error - did not get one")
+				assert.Equal(t, err.Error(), tt.wantErrString, "unexpected error message")
+			} else {
+				assert.Nil(t, err, "unexpected error")
+			}
+		})
+	}
 }
