+++ common/util_test.go
@@ -31,132 +31,11 @@ import (
 	"time"
 
 	"github.com/pborman/uuid"
-	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
-	commonpb "go.temporal.io/api/common/v1"
-	enumspb "go.temporal.io/api/enums/v1"
 	"go.temporal.io/api/serviceerror"
-	"google.golang.org/protobuf/types/known/durationpb"
 
 	"go.temporal.io/server/common/dynamicconfig"
 )
-
-func TestValidateRetryPolicy(t *testing.T) {
-	testCases := []struct {
-		name          string
-		input         *commonpb.RetryPolicy
-		wantErr       bool
-		wantErrString string
-	}{
-		{
-			name:          "nil policy is okay",
-			input:         nil,
-			wantErr:       false,
-			wantErrString: "",
-		},
-		{
-			name: "maxAttempts is 1, coefficient < 1",
-			input: &commonpb.RetryPolicy{
-				BackoffCoefficient: 0.5,
-				MaximumAttempts:    1,
-			},
-			wantErr:       false,
-			wantErrString: "",
-		},
-		{
-			name: "initial interval negative",
-			input: &commonpb.RetryPolicy{
-				InitialInterval: durationpb.New(-22 * time.Second),
-			},
-			wantErr:       true,
-			wantErrString: "InitialInterval cannot be negative on retry policy.",
-		},
-		{
-			name: "coefficient < 1",
-			input: &commonpb.RetryPolicy{
-				BackoffCoefficient: 0.8,
-			},
-			wantErr:       true,
-			wantErrString: "BackoffCoefficient cannot be less than 1 on retry policy.",
-		},
-		{
-			name: "maximum interval in seconds is negative",
-			input: &commonpb.RetryPolicy{
-				BackoffCoefficient: 2.0,
-				MaximumInterval:    durationpb.New(-2 * time.Second),
-			},
-			wantErr:       true,
-			wantErrString: "MaximumInterval cannot be negative on retry policy.",
-		},
-		{
-			name: "maximum interval in less than initial interval",
-			input: &commonpb.RetryPolicy{
-				BackoffCoefficient: 2.0,
-				MaximumInterval:    durationpb.New(5 * time.Second),
-				InitialInterval:    durationpb.New(10 * time.Second),
-			},
-			wantErr:       true,
-			wantErrString: "MaximumInterval cannot be less than InitialInterval on retry policy.",
-		},
-		{
-			name: "maximum attempts negative",
-			input: &commonpb.RetryPolicy{
-				BackoffCoefficient: 2.0,
-				MaximumAttempts:    -3,
-			},
-			wantErr:       true,
-			wantErrString: "MaximumAttempts cannot be negative on retry policy.",
-		},
-		{
-			name: "timeout nonretryable error - valid type",
-			input: &commonpb.RetryPolicy{
-				BackoffCoefficient: 1,
-				NonRetryableErrorTypes: []string{
-					TimeoutFailureTypePrefix + enumspb.TIMEOUT_TYPE_START_TO_CLOSE.String(),
-					TimeoutFailureTypePrefix + enumspb.TIMEOUT_TYPE_SCHEDULE_TO_START.String(),
-					TimeoutFailureTypePrefix + enumspb.TIMEOUT_TYPE_SCHEDULE_TO_CLOSE.String(),
-					TimeoutFailureTypePrefix + enumspb.TIMEOUT_TYPE_HEARTBEAT.String(),
-				},
-			},
-			wantErr:       false,
-			wantErrString: "",
-		},
-		{
-			name: "timeout nonretryable error - unspecified type",
-			input: &commonpb.RetryPolicy{
-				BackoffCoefficient: 1,
-				NonRetryableErrorTypes: []string{
-					TimeoutFailureTypePrefix + enumspb.TIMEOUT_TYPE_UNSPECIFIED.String(),
-				},
-			},
-			wantErr:       true,
-			wantErrString: "Invalid timeout type value: Unspecified.",
-		},
-		{
-			name: "timeout nonretryable error - unknown type",
-			input: &commonpb.RetryPolicy{
-				BackoffCoefficient: 1,
-				NonRetryableErrorTypes: []string{
-					TimeoutFailureTypePrefix + "unknown",
-				},
-			},
-			wantErr:       true,
-			wantErrString: "Invalid timeout type value: unknown.",
-		},
-	}
-
-	for _, tt := range testCases {
-		t.Run(tt.name, func(t *testing.T) {
-			err := ValidateRetryPolicy(tt.input)
-			if tt.wantErr {
-				assert.NotNil(t, err, "expected error - did not get one")
-				assert.Equal(t, err.Error(), tt.wantErrString, "unexpected error message")
-			} else {
-				assert.Nil(t, err, "unexpected error")
-			}
-		})
-	}
-}
 
 func TestIsContextDeadlineExceededErr(t *testing.T) {
 	ctx, cancel := context.WithDeadline(context.Background(), time.Now().Add(-time.Millisecond))
